Simplify personalised params to suit test template. Preview or send email from one form
