+++ backoffice/src/Models/Email/EmailPreview.cs
@@ -1,0 +1,8 @@
+namespace Droits.Models.Email;
+
+public class EmailPreview
+{
+    public Guid Id { get; set; }
+    public string Body { get; set; }
+    public string Subject { get; set; }
+}
