+++ backoffice/src/Views/Email/GetPreview.cshtml
@@ -1,3 +1,4 @@
+@using GovUk.Frontend.AspNetCore.TagHelpers
 @model Notify.Models.Responses.TemplatePreviewResponse
 
 @{
@@ -46,4 +47,3 @@
         }
     }
 }
-
