+++ backoffice/src/Views/Email/SendEmail.cshtml
@@ -1,4 +1,5 @@
-@model EmailForm
+@using Microsoft.AspNetCore.Mvc.TagHelpers
+@model Droits.Models.Email.EmailForm
 @{
     ViewData["Title"] = "Email Successfully Sent";
 }
