+++ backoffice/src/Views/Email/templates/TestingDroitsv2.txt
@@ -1,0 +1,1 @@
+((body))
